moved and wrote tests for that method
